+++ packages/hardhat/package.json
@@ -38,7 +38,8 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "-": "^0.0.1",
+    "boj-types": "^1.0.0",
+    "boj-utils": "^1.0.0",
     "@iarna/toml": "^2.2.5",
     "@nomicfoundation/hardhat-toolbox-viem": "^1.0.0",
     "@nomicfoundation/hardhat-viem": "1.0.2",
