+++ packages/hardhat/scripts/getVerifyCommand.ts
@@ -1,7 +1,8 @@
 import dotenv from "dotenv";
-import { delay, getContract, getDecryptedValue } from "../utils/utils";
 dotenv.config({ path: "../.env" });
 import hre from "hardhat";
+import { readDeploymentData } from "./saveDeploy";
+import { lchown } from "fs";
 
 const params = {
   account: {
@@ -11,18 +12,32 @@ const params = {
 };
 
 async function main() {
-  const privateToken = await getContract("PrivateToken");
-  const allTransferVerifier = await getContract("TransferVerify");
-  const allWithdrawVerifier = await getContract("WithdrawVerify");
-  const lockVerifier = await getContract(
-    "contracts/lock/plonk_vk.sol:UltraVerifier"
+  const getContractAt = hre.viem.getContractAt
+  const { data: privateTokenData } = readDeploymentData("PrivateToken");
+  const { data: transferVerifyData } = readDeploymentData("TransferVerify");
+  const { data: withdrawVerifyData } = readDeploymentData("WithdrawVerify");
+  const { data: lockData } = readDeploymentData("contracts/lock/plonk_vk.sol:UltraVerifier");
+  const { data: funTokenData } = readDeploymentData("FunToken");
+  const { data: processTransferData } = readDeploymentData("contracts/process_pending_transfers/plonk_vk.sol:UltraVerifier");
+  const { data: processDepositData } = readDeploymentData("contracts/process_pending_deposits/plonk_vk.sol:UltraVerifier");
+
+  const network = hre.network.name;
+
+  let privateToken = await getContractAt(
+    "PrivateToken",
+    privateTokenData[network].address
   );
-  const token = await getContract("FunToken");
-  const processDepositVerifier = await getContract(
-    "contracts/process_pending_deposits/plonk_vk.sol:UltraVerifier"
+  const allTransferVerifier = await getContractAt("TransferVerify", transferVerifyData[network].address);
+  const allWithdrawVerifier = await getContractAt("WithdrawVerify", withdrawVerifyData[network].address);
+  const lockVerifier = await getContractAt(
+    "contracts/lock/plonk_vk.sol:UltraVerifier", lockData[network].address
   );
-  const processTransferVerifier = await getContract(
-    "contracts/process_pending_transfers/plonk_vk.sol:UltraVerifier"
+  const token = await getContractAt("FunToken", funTokenData[network].address);
+  const processDepositVerifier = await getContractAt(
+    "contracts/process_pending_deposits/plonk_vk.sol:UltraVerifier", processDepositData[network].address
+  );
+  const processTransferVerifier = await getContractAt(
+    "contracts/process_pending_transfers/plonk_vk.sol:UltraVerifier", processTransferData[network].address
   );
 
   const decimals = await token.read.decimals();
