+++ packages/types/package.json
@@ -6,6 +6,6 @@
   "type": "module",
   "license": "MIT",
   "scripts": {
-    "build": "tsc"
+    "build": "yarn clean && tsc -b"
   }
 }
