+++ packages/utils/src/constants.ts
@@ -1,0 +1,35 @@
+import { Address } from "viem";
+
+export const account2 = {
+    packedPublicKey:
+        "0x2713781b8a47737cc765d8b1e134c84c6c4764e8d4dbfc30e93a84606fb81d9d" as `0x${string}`,
+    privateKey:
+        "0x01e6ca61e1d33b85021d0ab4e651d069660c3c7521a3fbafe58dfa95798e884f" as `0x${string}`,
+};
+// old account 2 ("0x0c07999c15d406bc08d7f3f31f62cedbc89ebf3a53ff4d3bf7e2d0dda9314904");
+export const account1 = {
+    packedPublicKey:
+        "0xdc9f9fdb746d0f07b004cc4316e3495a58570b90661499f8a6a6696ff4156baa" as `0x${string}`,
+    privateKey:
+        "0x0510bae26a9b59ebad67a4324c944b1910a778e8481d7f08ddba6bcd2b94b2c4" as `0x${string}`,
+};
+export const processFeeRecipient =
+    "0xdead940f35737EDb9a9Ad2bB938A955F9b7892e3" as `0x${string}`;
+
+export const withdrawAddress =
+    "0xbead940f35737EDb9a9Ad2bB938A955F9b7892e4" as Address;
+
+export const depositAmount = BigInt(10 * 10 ** 18);
+export const depositProcessFee = 1;
+export const transferProcessFee = 2;
+export const transferRelayFee = 2;
+export const transferAmount = 5;
+
+export const BJJ_PRIME =
+    21888242871839275222246405745257275088548364400416034343698204186575808495617n;
+
+export const MAX_TXS_TO_PROCESS = 4;
+
+// 168986485046885582825082387270879151100288537211746581237924789162159767775n
+export const random =
+    "0x005fa49918e16a3c3f4d42071a137c8a818964701c63604244ac2746b95e00df";
