--- packages/utils/yarn-error.log
@@ -1,102 +1,0 @@
-Arguments: 
-  /home/josh/.nvm/versions/node/v18.17.1/bin/node /home/josh/.nvm/versions/node/v18.17.1/bin/yarn add -D fflate @noir-lang/acvm
-
-PATH: 
-  /home/josh/.bun/bin:/home/josh/.rbenv/shims:/home/josh/.nix-profile/bin:/home/josh/.bun/bin:/home/josh/.rbenv/shims:/home/josh/.rbenv/bin:/home/josh/.nvm/versions/node/v18.17.1/bin:/home/josh/.nix-profile/bin:/home/josh/.nix-profile/bin:/home/josh/.cargo/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin:/snap/bin:/home/josh/.foundry/bin:/home/josh/.nargo/bin:/home/josh/.nargo/bin:/home/josh/.yarn/bin:/home/josh/.foundry/bin:/home/josh/.nargo/bin:/home/josh/.nargo/bin:/home/josh/.yarn/bin
-
-Yarn version: 
-  1.22.19
-
-Node version: 
-  18.17.1
-
-Platform: 
-  linux x64
-
-Trace: 
-  Error: https://registry.yarnpkg.com/@noir-lang%2facvm: Not found
-      at params.callback [as _callback] (/home/josh/.nvm/versions/node/v18.17.1/lib/node_modules/yarn/lib/cli.js:66145:18)
-      at self.callback (/home/josh/.nvm/versions/node/v18.17.1/lib/node_modules/yarn/lib/cli.js:140890:22)
-      at Request.emit (node:events:514:28)
-      at Request.<anonymous> (/home/josh/.nvm/versions/node/v18.17.1/lib/node_modules/yarn/lib/cli.js:141862:10)
-      at Request.emit (node:events:514:28)
-      at IncomingMessage.<anonymous> (/home/josh/.nvm/versions/node/v18.17.1/lib/node_modules/yarn/lib/cli.js:141784:12)
-      at Object.onceWrapper (node:events:628:28)
-      at IncomingMessage.emit (node:events:526:35)
-      at endReadableNT (node:internal/streams/readable:1359:12)
-      at process.processTicksAndRejections (node:internal/process/task_queues:82:21)
-
-npm manifest: 
-  {
-    "name": "utils",
-    "version": "1.0.0",
-    "description": "",
-    "type": "module",
-    "main": "test.js",
-    "scripts": {
-      "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "author": "",
-    "license": "ISC",
-    "devDependencies": {
-      "ethers": "^6.8.0"
-    }
-  }
-
-yarn manifest: 
-  No manifest
-
-Lockfile: 
-  # THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
-  # yarn lockfile v1
-  
-  
-  "@adraffy/ens-normalize@1.10.0":
-    version "1.10.0"
-    resolved "https://registry.yarnpkg.com/@adraffy/ens-normalize/-/ens-normalize-1.10.0.tgz#d2a39395c587e092d77cbbc80acf956a54f38bf7"
-    integrity sha512-nA9XHtlAkYfJxY7bce8DcN7eKxWWCWkU+1GR9d+U6MbNpfwQp8TI7vqOsBsMcHoT4mBu2kypKoSKnghEzOOq5Q==
-  
-  "@noble/curves@1.2.0":
-    version "1.2.0"
-    resolved "https://registry.yarnpkg.com/@noble/curves/-/curves-1.2.0.tgz#92d7e12e4e49b23105a2555c6984d41733d65c35"
-    integrity sha512-oYclrNgRaM9SsBUBVbb8M6DTV7ZHRTKugureoYEncY5c65HOmRzvSiTE3y5CYaPYJA/GVkrhXEoF0M3Ya9PMnw==
-    dependencies:
-      "@noble/hashes" "1.3.2"
-  
-  "@noble/hashes@1.3.2":
-    version "1.3.2"
-    resolved "https://registry.yarnpkg.com/@noble/hashes/-/hashes-1.3.2.tgz#6f26dbc8fbc7205873ce3cee2f690eba0d421b39"
-    integrity sha512-MVC8EAQp7MvEcm30KWENFjgR+Mkmf+D189XJTkFIlwohU5hcBbn1ZkKq7KVTi2Hme3PMGF390DaL52beVrIihQ==
-  
-  "@types/node@18.15.13":
-    version "18.15.13"
-    resolved "https://registry.yarnpkg.com/@types/node/-/node-18.15.13.tgz#f64277c341150c979e42b00e4ac289290c9df469"
-    integrity sha512-N+0kuo9KgrUQ1Sn/ifDXsvg0TTleP7rIy4zOBGECxAljqvqfqpTfzx0Q1NUedOixRMBfe2Whhb056a42cWs26Q==
-  
-  aes-js@4.0.0-beta.5:
-    version "4.0.0-beta.5"
-    resolved "https://registry.yarnpkg.com/aes-js/-/aes-js-4.0.0-beta.5.tgz#8d2452c52adedebc3a3e28465d858c11ca315873"
-    integrity sha512-G965FqalsNyrPqgEGON7nIx1e/OVENSgiEIzyC63haUMuvNnwIgIjMs52hlTCKhkBny7A2ORNlfY9Zu+jmGk1Q==
-  
-  ethers@^6.8.0:
-    version "6.8.0"
-    resolved "https://registry.yarnpkg.com/ethers/-/ethers-6.8.0.tgz#0a26f57e96fd697cefcfcef464e0c325689d1daf"
-    integrity sha512-zrFbmQRlraM+cU5mE4CZTLBurZTs2gdp2ld0nG/f3ecBK+x6lZ69KSxBqZ4NjclxwfTxl5LeNufcBbMsTdY53Q==
-    dependencies:
-      "@adraffy/ens-normalize" "1.10.0"
-      "@noble/curves" "1.2.0"
-      "@noble/hashes" "1.3.2"
-      "@types/node" "18.15.13"
-      aes-js "4.0.0-beta.5"
-      tslib "2.4.0"
-      ws "8.5.0"
-  
-  tslib@2.4.0:
-    version "2.4.0"
-    resolved "https://registry.yarnpkg.com/tslib/-/tslib-2.4.0.tgz#7cecaa7f073ce680a05847aa77be941098f36dc3"
-    integrity sha512-d6xOpEDfsi2CZVlPQzGeux8XMwLT9hssAsaPYExaQMuYskwb+x1x7J371tWlbBdWHroy99KnVB6qIkUbs5X3UQ==
-  
-  ws@8.5.0:
-    version "8.5.0"
-    resolved "https://registry.yarnpkg.com/ws/-/ws-8.5.0.tgz#bfb4be96600757fe5382de12c670dab984a1ed4f"
-    integrity sha512-BWX0SWVgLPzYwF8lTzEy1egjhS4S4OEAHfsO8o65WOVsrnSRGaSiUaa9e0ggGlkMTtBlmOpEXiie9RUcBO86qg==
